Fixed a bug with MulticastClient receiving unwanted messages.
A threading issue was resolved which caused the MulticastClient to receive one more message than it was supposed to, and sending it out to all the callbacks that were subscribed to the client.

The MulticastClientTest was updated to be more robust and consistent, and has passed all tests.

Updated documentation and comments.